Move assert_batches_eq! macros to test_utils.rs (#746)
* Move assert_batches_eq! macros to test_utils.rs

* port test
